check last name value is digits rather than number of '.'
New custom builds of TD have a daily number appended to the end of the name so search for 'TouchDesigner' strings in the registry with exactly 2 periods (.) won't append those entries. Instead of looking for 2 periods check if the last entry is digits and assume it relates to a valid build.